backend/requirements-test.txt
# ─── Test runner ──────────────────────────────────────────────────────────────
pytest>=9.0.3
pytest-asyncio>=1.3.0
pytest-cov>=6.0.0

# ─── HTTP client (used by FastAPI TestClient internally) ──────────────────────
httpx>=0.28.1
anyio>=4.13.0

# ─── SQLite driver (stdlib on Python 3, no install needed) ───────────────────
# pysqlite3 is bundled with Python — no extra package required

# ─── App dependencies that the tests import directly ─────────────────────────
fastapi>=0.136.0
sqlalchemy>=2.0.49
pydantic>=2.13.2
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.5.0
bcrypt>=4.0.1